quant_research_lab/requirements.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
requests>=2.31.0
python-dotenv>=1.0.0

# Machine Learning & Feature Engineering
lightgbm>=4.0.0
scikit-learn>=1.3.0
ta>=0.10.2

# Scheduling & Time
APScheduler>=3.10.4

# API & Data
ta-lib>=0.4.24

# Jupyter & Visualization
jupyter>=1.0.0
matplotlib>=3.7.0
plotly>=5.14.0
seaborn>=0.12.0

# Advanced ML Interpretation
shap>=0.42.0
optuna>=3.14.0

# Backtesting
vectorbt>=0.25.0

# Code Quality
pylint>=2.17.0

# Optional: Database
# sqlite3 (built-in with Python)